+++ packages/editor/src/components/provider/disable-non-page-content-blocks.js
@@ -4,13 +4,14 @@
 import { useSelect, useDispatch } from '@wordpress/data';
 import { store as blockEditorStore } from '@wordpress/block-editor';
 import { useEffect } from '@wordpress/element';
+import { applyFilters } from '@wordpress/hooks';
 
-const CONTENT_ONLY_BLOCKS = [
-	'core/post-content',
-	'core/post-featured-image',
+const CONTENT_ONLY_BLOCKS = applyFilters( 'editor.postContentBlockTypes', [
 	'core/post-title',
+	'core/post-featured-image',
+	'core/post-content',
 	'core/template-part',
-];
+] );
 
 /**
  * Component that when rendered, makes it so that the site editor allows only